Rename index.ts -> app-ui-state-manager
